Added ability to control which listener is used in the Contact field of SIP REGISTER.
Moved the parameters for SIP REGISTER to a struct so we can add with impunity in the future without adding endless parameters to functions.

Fixed issue with non-connection based transactions, eg REGISTER, and multi-homed systems. In particular when authentication is involved.

Some general code clean ups in the SIP handlers.