Added ability for a plug in codec to get access to the OPAL/PTLib logging system, without including PTLib itself, by use of a call back function.Changes to the internal structures so can actually make a call to ones self. Why would you possibly want to do that? Well, some server applications can end up doing it, usually with different destination user names.
Includes some code tidying so that when a call-id or a dialog tag is created a single function does the creation, allowing for experimenting with algorithms on something other than the rather long GUID to be used, without changing code in twelve places.Changes to the internal structures so can actually make a call to ones self. Why would you possibly want to do that? Well, some server applications can end up doing it, usually with different destination user names.