Make sure if unsubscribing/unregistering, it returns true if it already unsubscribed/unregistered.
Unfortunately fix for the original reported problem is too much for stable "bug fix". This is to get around the limitation that you cannot start an unregister (for example) while a register is currently in progress. This is dictated by the protocol, not implementation, and the implementation is going to punt to the application.