Commit test program to measure the response time of a PSyncPoint.
Code uses a STL list, so is an example of constructing and populating a STL list.

On linux, the worst response time was 35 microseconds.

On Windows, the response time is waiting someone to compile it and report back.

Full doxygen comments are included, so if you have more data, please add it to main.h