Back ports: Fixed SIP 302 redirect response so redirects to correct place.
Fixed issue with deadlock in INVITE with replaces clause, when sending BYE and getting 200 OK for the BYE, do not need to do anything with the response, and most importantly do not need to lock the SIPConnection.

Fixed missing NOTIFY when we send a 180/183 response to move to "Early" state.