Fixed some OpenPhone video issues:
The global opening of video input deviceo prevented video test in options dialog when have camera selected. Now opens only opens the camera when a video call is started or received.

Fixed video test so if close the window, it sends a message back to the dialog to reset the video test button from "stop" back to "test".