Fixed bug #483 "HandleUnreachable handles RX channel as TX channel"
If the ICMP "unreachable" error appears on the recvfrom() in the RTP processing, it kills the receive RTP thread, preventing the time out from occurring (or not) in that direction of the media flow. It is possible for media to still be received, in which cans to call continues, or we time out and kill the call. The thread exiting means the time out never occurs so we never kill the call.